Split TmuxBackend into focused extension files
Keep TmuxBackend as a thin facade with shared state and callbacks, and
move each cockpit concern into its own file so terminal bugs no longer
share a 1.6k-line review. Public API, paste timing, and alt-buffer
latch semantics are unchanged.

Crow-Session: 19DA2F69-BA78-4F6C-8A9E-B8AD0AC6A067
