+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- CROW-1222 — Split `TmuxBackend.swift` (~1.6k lines) into a thin facade plus focused `extension TmuxBackend` files (`TmuxServerLifecycle`, `TmuxWindowLifecycle`, `TmuxInput`, `TmuxScrollbackHealth`, `TmuxCockpitReaper`, `TmuxReadinessWatch`, `TmuxManagerExitMonitor`). Call sites keep `TmuxBackend.shared.*`; paste-Enter settle, alt-buffer latch, and the 50k history floor are unchanged.
+
 - CROW-1215 — Claude Code **≥ 2.1.259** `--permission-prompts none` (deny anything that would prompt while auto still classifies; Containment Escape still applies) was evaluated against ADR 0015 and **declined**. It is print-mode only (`claude -p`; CLI reference: "Set who answers permission prompts in print mode"). Crow launches the interactive TUI in a tmux PTY, so the flag cannot convert CROW-1176's extra-workdir auto-mode Read stalls into denials for `JobScheduler` / Manager. Even if it became TUI-legal, applying it on the Manager would deny `$HOME` / `$TMPDIR` Reads that today only stall. Crow still emits `--permission-mode auto` and nothing else; launch tests pin that no Claude path emits `--permission-prompts`. [ADR 0015](docs/adr/0015-harness-capability-tiers.md) records the eval. Re-check only if the flag becomes TUI-legal — then `.job` only, never Manager.
 
 - CROW-1176 — Claude `--permission-mode auto` is **no longer stall-free**. Claude Code ≥ 2.1.257 prompts once before the first file Read outside the working directories (launch cwd + `--add-dir` + `permissions.additionalDirectories`). Crow still emits `--permission-mode auto` for Manager / `.job` / (when enabled) coder-view Auto and **does not** paper over that prompt: `--dangerously-skip-permissions` / `bypassPermissions` would also drop the same-release Containment Escape rule (brittle-reject), `permissions.blockReadsOutsideWorkingDirectories` only *blocks* extra-workdir Reads, and a blanket `--add-dir` of `$HOME` / `$TMPDIR` / sibling checkouts is the too-permissive fake. Review clones under `{devRoot}/crow-reviews/` are the likely first hit. Matrix Auto-permission cell is now ⚠️; [ADR 0015](docs/adr/0015-harness-capability-tiers.md) records the eval. A live Crow-tmux PTY pass on ≥ 2.1.257 remains a human re-check (eval binary was 2.1.232).
