+++ Packages/CrowTerminal/Sources/CrowTerminal/TmuxBackend.swift
@@ -14,13 +14,20 @@ protocol CockpitSessionStarter {
 /// Crow-app-wide singleton that owns the tmux server backing all
 /// `SessionTerminal.backend == .tmux` rows.
 ///
-/// Responsibilities:
-///   - Lazily start a per-app tmux server with the bundled `crow-tmux.conf`.
-///   - Map terminal UUIDs to tmux window indices.
-///   - Drive `select-window` / `new-window` / `kill-window` / `paste-buffer`
-///     in response to UI events from the rest of the app.
-///   - Track readiness via `SentinelWaiter` (replaces the historical 5s
-///     sleep from the old per-terminal renderer path).
+/// Facade (CROW-1222): `configure`, shared state, callbacks, and the public
+/// methods callers already use. Each concern lives in its own file as an
+/// `extension TmuxBackend` so call sites keep `TmuxBackend.shared.*` and
+/// Swift access stays module-internal — no second singleton, no public-API
+/// move.
+///
+///   - `TmuxServerLifecycle` — ensure / resurrect, conf reconcile, reload
+///   - `TmuxWindowLifecycle` — register / adopt / destroy / kill / list
+///   - `TmuxInput` — paste, search, prompt-nav, select-all, cwd
+///   - `TmuxScrollbackHealth` — CROW-804 / CROW-1023 / #822
+///   - `TmuxCockpitReaper` — unbound reap + orphan reconcile (not the
+///     legacy PID-socket `TmuxOrphanReaper`)
+///   - `TmuxReadinessWatch` — sentinel wait, retry, diagnostics
+///   - `TmuxManagerExitMonitor` — Manager foreground-command poll (#558)
 ///
 /// Thread-safety: `@MainActor` — window/binding mutations are confined to the
 /// main actor. (The daemon renders terminals in-browser over the `/terminal`
@@ -63,19 +70,23 @@ public final class TmuxBackend {
     /// the next tmux command.
     public var onServerLost: (() -> Void)?
 
-    // MARK: - Internal state
+    // MARK: - Shared state (CROW-1222)
+    //
+    // Module-internal so the `extension TmuxBackend` files can reach it.
+    // Swift `private` is file-scoped; these are not public and this is not a
+    // second singleton.
 
     /// Created on first use of the backend. Survives until app exit (or a
     /// `shutdown()` call from the watchdog flow in PROD #5).
-    private var controller: TmuxController?
+    var controller: TmuxController?
 
     /// UUID → tmux window index for tabs registered with us.
-    private var bindings: [UUID: Int] = [:]
+    var bindings: [UUID: Int] = [:]
 
     /// Agent-window indices seen orphaned on the previous reconcile pass — the
     /// one-pass grace so a window created mid-`new-terminal` (before its binding
     /// lands) is never reaped (CROW-581).
-    private var orphanGraceWindows: Set<Int> = []
+    var orphanGraceWindows: Set<Int> = []
 
     /// Window indices observed in the alternate buffer (`#{alternate_on}==1`) at
     /// least once, latched sticky (CROW-1023). This is the ground truth for the
@@ -95,7 +106,7 @@ public final class TmuxBackend {
     /// stays `alt_on=0`. Pruned to live windows on each read, and cleared per
     /// index at (re)registration, so a recycled tmux index never inherits a
     /// stale observation.
-    private var observedAltBufferWindows: Set<Int> = []
+    var observedAltBufferWindows: Set<Int> = []
 
     /// Terminal whose tmux window is currently selected, so `makeActive` can
     /// skip a redundant `select-window`. `makeActive` is called repeatedly for
@@ -103,26 +114,26 @@ public final class TmuxBackend {
     /// without this each call shells out another run-loop-pumping subprocess
     /// (review nit on #336). Keyed by UUID, not window index — tmux can reuse
     /// a freed index for a new window, and a UUID never collides that way.
-    private(set) var activeTerminalID: UUID?
+    var activeTerminalID: UUID?
 
     /// UUID → per-terminal sentinel path. Cleared on destroy.
-    private var sentinels: [UUID: String] = [:]
+    var sentinels: [UUID: String] = [:]
 
     /// UUID → per-terminal wrapper-log path. Populated alongside `sentinels`
     /// so `captureDiagnostics(id:)` can read it back on `.timedOut`. Cleared
     /// on destroy. Issue #256.
-    private var wrapperLogs: [UUID: String] = [:]
+    var wrapperLogs: [UUID: String] = [:]
 
     /// UUID → in-flight readiness watch Tasks (the 10s progress beacon and
     /// the waiter). `destroyTerminal` cancels these so they don't fire
     /// `onReadinessChanged` for a tab the user just closed. Issue #282.
-    private var readinessTasks: [UUID: [Task<Void, Never>]] = [:]
+    var readinessTasks: [UUID: [Task<Void, Never>]] = [:]
 
     /// Poll loop watching the Manager window's foreground command so the
     /// "Manager process exited" banner reappears under the shared xterm.js
     /// attach client (#558). At most one runs at a time; re-armed by
     /// `SessionService` on launch / restart. See `startManagerExitMonitor`.
-    private var managerExitMonitor: Task<Void, Never>?
+    var managerExitMonitor: Task<Void, Never>?
 
     /// Public for test isolation. Production callers use `.shared`.
     public init() {}
@@ -230,876 +241,6 @@ public final class TmuxBackend {
         wrapperLogs.removeAll()
     }
 
-    // MARK: - Per-terminal API
-
-    /// Create a new tmux window for `id`. If the cockpit session doesn't
-    /// exist yet, starts it. Returns the binding so callers can persist
-    /// it on the `SessionTerminal` row.
-    ///
-    /// `agentSurface` selects this window's scroll model (ADR-0013): a
-    /// repainting agent TUI gets `alternate-screen on`, while a plain shell
-    /// keeps the global `off` and the unified 50k scrollback. Callers should
-    /// pass `SessionTerminal.isAgentSurface(session:)` rather than a hand-rolled
-    /// test — in particular `isManaged` ALONE is wrong, because the Manager's
-    /// terminal is built without that flag yet still runs an agent.
-    ///
-    /// `usesAlternateScreen` is the per-agent capability (CROW-1008 / CROW-1010).
-    /// `true` (Claude Code) relies on the alt buffer to kill sediment. `false`
-    /// (Cursor, and any inline renderer) still gets `alternate-screen on` so
-    /// `list-terminals` classifies the window as an agent surface, but keeps
-    /// the unified 50k history — its transcript is legitimate native scrollback
-    /// and the wheel at a non-mouse-tracking prompt scrolls it (#850). The
-    /// CROW-1008 `history-limit 0` clamp is retracted: it deleted Cursor's only
-    /// scroll path. The flag is forwarded to the client as
-    /// `uses_alternate_screen` so `applySurfaceScrollback` caps xterm only for
-    /// true alt-buffer agents.
-    @discardableResult
-    public func registerTerminal(
-        id: UUID,
-        name: String,
-        cwd: String,
-        command: String?,
-        trackReadiness: Bool,
-        agentKind: AgentKind? = nil,
-        agentSurface: Bool = false,
-        usesAlternateScreen _: Bool = true,
-        extraEnv: [String: String] = [:],
-        newWindowTimeout: TimeInterval = TmuxController.defaultTimeout
-    ) throws -> TmuxBinding {
-        precondition(!tmuxBinary.isEmpty, "TmuxBackend.configure(...) must be called first")
-        let ctrl: TmuxController
-        do {
-            ctrl = try ensureRunningServer()
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-
-        // Each window gets its own sentinel path so concurrent terminals
-        // don't race on the same file.
-        let sentinelPath = sentinelPath(for: id)
-        try? FileManager.default.removeItem(atPath: sentinelPath)
-        sentinels[id] = sentinelPath
-
-        // Per-terminal wrapper log. The bundled shell wrapper writes stage
-        // breadcrumbs here so `captureDiagnostics(id:)` can include them in
-        // the .timedOut bundle (issue #256).
-        let wrapperLog = wrapperLogPath(for: id)
-        try? FileManager.default.removeItem(atPath: wrapperLog)
-        wrapperLogs[id] = wrapperLog
-
-        // Shell wrapper does the readiness markers + sources user's shell
-        // config. Each tmux window's child process *is* the wrapper.
-        guard let wrapperURL = BundledResources.shellWrapperScriptURL else {
-            throw TmuxBackendError.bundledResourceMissing("crow-shell-wrapper.sh")
-        }
-        let wrapperPath = wrapperURL.path
-
-        var env = [
-            "CROW_SENTINEL": sentinelPath,
-            "CROW_WRAPPER_LOG": wrapperLog,
-        ]
-        // Caller-supplied vars (e.g. CROW_ARTIFACTS_DIR / CROW_SESSION_ID).
-        // Merged first so the built-ins below always win on any collision.
-        for (key, value) in extraEnv { env[key] = value }
-        if let agentKind {
-            for (key, value) in CrowAttribution.environmentEntries(for: agentKind) {
-                env[key] = value
-            }
-        }
-        if !cwd.isEmpty { env["PWD"] = cwd }
-
-        // CROW-487: hand the per-devroot bin dir to the wrapper so it can
-        // prepend it to PATH *after* user rc sourcing — that's the only
-        // insertion point that survives `export PATH=…` in `.zshrc`. We also
-        // seed the window's PATH directly so non-rc shells (fish, the
-        // unknown-shell fallback branch of the wrapper, processes that
-        // bypass the wrapper entirely) still find the symlink farm.
-        if !crowBinDir.isEmpty {
-            env["CROW_BIN_DIR"] = crowBinDir
-            env["PATH"] = "\(crowBinDir):\(ShellEnvironment.shared.resolvedPATH)"
-        }
-
-        // CROW-1010: do not clamp `history-limit` for inline agents. Cursor's
-        // history is a clean transcript and the local-viewport wheel (#850) is
-        // its only scroll path; `history-limit 0` made that wheel a no-op.
-        // Alt-buffer agents (Claude Code) don't need a clamp either — the alt
-        // buffer has no history. `usesAlternateScreen` is still accepted so
-        // callers keep passing the capability; list-terminals reads it from
-        // AgentRegistry and forwards `uses_alternate_screen` to the client.
-
-        let windowIndex = try ctrl.newWindow(
-            name: name,
-            cwd: cwd.isEmpty ? nil : cwd,
-            env: env,
-            command: wrapperPath,
-            timeout: newWindowTimeout
-        )
-        bindings[id] = windowIndex
-        // CROW-1023: a freshly created window has not entered the alt buffer
-        // yet, so forget any observation a prior tenant of this index left
-        // behind (tmux reuses freed indices). Detection re-latches on the next
-        // `list-windows` read once this window actually enters the alt screen.
-        observedAltBufferWindows.remove(windowIndex)
-
-        // Hand agent-TUI windows their own viewport BEFORE the launch command
-        // below is pasted, so an agent that *does* request smcup enters the alt
-        // buffer on its very first repaint and never deposits a frame into the
-        // shared history (#822). For inline renderers the option is inert but
-        // still the source of truth for `agent_surface` on list-terminals.
-        if agentSurface {
-            enableAlternateScreen(index: windowIndex)
-        }
-
-        if trackReadiness {
-            startReadinessWatch(id: id, sentinelPath: sentinelPath)
-        }
-
-        // If the caller supplied an initial command (e.g. `claude --continue`),
-        // route it through the buffer-paste path — same as PROD #3.
-        if let command, !command.isEmpty {
-            try sendText(id: id, text: command + "\n")
-        }
-
-        return TmuxBinding(
-            socketPath: ctrl.socketPath,
-            sessionName: ctrl.sessionName,
-            windowIndex: windowIndex
-        )
-    }
-
-    /// Re-bind a terminal to a window that already exists in the live tmux
-    /// server (e.g. on app restart with a long-lived session). No new
-    /// window is created.
-    public func adoptTerminal(id: UUID, binding: TmuxBinding, trackReadiness: Bool) throws {
-        let (ctrl, serverWasResurrected) = try ensureRunningServerReportingResurrection()
-        guard ctrl.socketPath == binding.socketPath, ctrl.sessionName == binding.sessionName else {
-            throw TmuxBackendError.bindingMismatch(
-                expected: binding.socketPath + ":" + binding.sessionName,
-                actual: ctrl.socketPath + ":" + ctrl.sessionName
-            )
-        }
-        let liveIndices = try ctrl.listWindowIndices()
-        guard liveIndices.contains(binding.windowIndex) else {
-            // A binding exists but the server had to be respawned from
-            // scratch: every window is gone, not just this one — the server
-            // crashed (or the machine rebooted) since the binding was made.
-            // Distinguish that from a single closed window (#588).
-            throw serverWasResurrected
-                ? TmuxBackendError.serverCrashed
-                : TmuxBackendError.windowNotFound(binding.windowIndex)
-        }
-        bindings[id] = binding.windowIndex
-        // No sentinel re-fire on adoption — the wrapper's precmd already
-        // touched the file when the original window was created.
-        let sentinelPath = sentinelPath(for: id)
-        sentinels[id] = sentinelPath
-        wrapperLogs[id] = wrapperLogPath(for: id)
-        if trackReadiness, FileManager.default.fileExists(atPath: sentinelPath) {
-            onReadinessChanged?(id, .shellReady)
-        } else if trackReadiness {
-            startReadinessWatch(id: id, sentinelPath: sentinelPath)
-        }
-    }
-
-    /// Bring `id`'s window into focus. Called by the UI when the user
-    /// switches tabs.
-    ///
-    /// The `select-window` shell-out runs synchronously on the main actor so tab
-    /// switches stay strictly serialized — the dedup guard, the `select-window`,
-    /// and the `activeTerminalID` update are atomic w.r.t. other switches, giving
-    /// last-switch-wins. This is safe even mid-window-open-animation because
-    /// `TmuxController.run` no longer pumps a nested run loop while waiting
-    /// (#653): the main thread blocks on the child without re-entering the
-    /// in-flight CoreAnimation commit that used to SIGSEGV. Deliberately NOT moved
-    /// off the main actor — doing so let two in-flight switches race and could
-    /// leave tmux focused on the wrong window until the next `makeActive`
-    /// (PR #658 review).
-    public func makeActive(id: UUID) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        // Already the selected window — skip the redundant `select-window`
-        // subprocess (see `activeTerminalID`).
-        if id == activeTerminalID { return }
-        let start = Date()
-        do {
-            try ensureRunningServer().selectWindow(index: windowIndex)
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-        // Record only after a successful switch — a failed `select-window`
-        // must not suppress the next attempt.
-        activeTerminalID = id
-        let elapsedMS = Int((Date().timeIntervalSince(start)) * 1000)
-        // Operator-greppable: `[CrowTelemetry tmux:tab_switch_ms=…]`. Easy
-        // to graph from logs today; trivially re-routed to a real metrics
-        // pipeline once one exists.
-        CrowLog.info("[CrowTelemetry tmux:tab_switch_ms=\(elapsedMS) terminal=\(id)]")
-    }
-
-    /// Settle time between `paste-buffer` and the submitting `Enter`.
-    ///
-    /// Bracketed-paste TUIs (Claude Code, Cursor's `agent`) need the
-    /// `\e[201~` bracket-end to finish before Enter arrives. 50ms was enough
-    /// for short Claude auto-respond lines (#272) but large Cursor pastes
-    /// (multi-KB Manager / job prompts) still race: the agent starts working
-    /// while sticky text remains in the composer, looking like a double paste
-    /// (#631). 200ms clears the composer for those payloads without a second
-    /// Enter (which would re-submit leftover text).
-    public static let pasteEnterSettleDelay: TimeInterval = 0.2
-
-    /// Send text to `id`'s window via the buffer-paste path. Works for
-    /// arbitrary-size payloads (Phase 3 §3 finding: send-keys -l fails
-    /// on >10KB; load-buffer + paste-buffer scales to 50KB+ in 133ms).
-    ///
-    /// Quirk: agent TUIs enable bracketed-paste mode, which wraps
-    /// `paste-buffer` output in `\e[200~…\e[201~`. A trailing `\n` inside the
-    /// bracket is treated as literal text, not as Enter — so prompts that
-    /// rely on `\n` to submit (quick actions, auto-respond) get pasted but
-    /// never submitted (#264). Strip the trailing newline before pasting and
-    /// deliver a separate `Enter` via `send-keys` afterwards.
-    ///
-    /// `pasteEnterSettleDelay` between the paste and the Enter keystroke
-    /// gives the TUI time to process the bracket-end sequence (`\e[201~`).
-    /// Without this, Enter can arrive early: Claude may drop it entirely
-    /// (#272); Cursor may submit but leave the prompt sitting in the input
-    /// box (#631).
-    ///
-    /// We also pre-cancel copy-mode on the pane before any delivery (#486).
-    /// The bundled `crow-tmux.conf` keeps `mouse on` so wheel scrollback
-    /// works (#452), but the default `WheelUpPane` puts the pane into
-    /// copy-mode, where both `paste-buffer` and `send-keys Enter` are
-    /// silently consumed by copy-mode key bindings instead of reaching the
-    /// underlying shell. Without the cancel, every programmatic send into
-    /// a pane the user has scrolled (Manager paste, auto-respond, quick
-    /// actions, bare-Enter submits) is dropped.
-    public func sendText(id: UUID, text: String) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            let endsWithNewline = text.hasSuffix("\n")
-            let payload = endsWithNewline ? String(text.dropLast()) : text
-
-            // Cancel copy-mode if the user scrolled the pane into it before
-            // we deliver anything. Covers both the paste-buffer path (which
-            // is a no-op in copy-mode) and the bare-Enter path (where
-            // `send-keys Enter` would otherwise hit the copy-mode key table
-            // — default emacs `copy-selection-and-cancel`, vi `cancel` —
-            // exiting copy-mode without delivering a CR to the shell (#486).
-            try ctrl.cancelCopyModeIfActive(target: target)
-
-            var didPaste = false
-            if !payload.isEmpty {
-                let bufferName = "crow-\(id.uuidString)"
-                try ctrl.loadBufferFromStdin(name: bufferName, data: Data(payload.utf8))
-                defer { ctrl.deleteBuffer(name: bufferName) }
-                try ctrl.pasteBuffer(name: bufferName, target: target)
-                didPaste = true
-            }
-            if endsWithNewline {
-                // Give the TUI time to process the paste bracket-end before
-                // the Enter key arrives. Only needed when we actually pasted
-                // content — a bare "\n" (Enter-only) needs no delay.
-                if didPaste {
-                    Thread.sleep(forTimeInterval: Self.pasteEnterSettleDelay)
-                }
-                try ctrl.sendKeys(target: target, keys: ["Enter"])
-            }
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Drop the scrollback buffer for terminal `id` via `tmux clear-history`.
-    /// On-screen rows survive — only the off-screen history is wiped — matching
-    /// what macOS Terminal "Clear" and iTerm2 "Clear Buffer" do. Surfaced from
-    /// the terminal context menu.
-    public func clearHistory(id: UUID) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            _ = try ctrl.run(["clear-history", "-t", target])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Enter tmux copy-mode and select the entire scrollback for terminal
-    /// `id` — the "Select All" equivalent for a terminal pane. Surfaced from
-    /// the terminal context menu. After this, Copy
-    /// (or Cmd+C) writes the captured text to the macOS pasteboard via the
-    /// existing `copy-pipe-no-clear` binding.
-    public func selectAll(id: UUID) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            // -H makes copy-mode enter without scrolling the screen first.
-            _ = try ctrl.run(["copy-mode", "-H", "-t", target])
-            try ctrl.sendKeys(target: target, keys: ["-X", "history-top"])
-            try ctrl.sendKeys(target: target, keys: ["-X", "begin-selection"])
-            try ctrl.sendKeys(target: target, keys: ["-X", "history-bottom"])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Read the live working directory of terminal `id`'s pane via
-    /// `tmux display-message -p -F '#{pane_current_path}'`. Used by
-    /// smart-detect `path:line` resolution (#471 gap 5) to honour the
-    /// pane's *current* cwd rather than the cockpit surface's static
-    /// `workingDirectory` (which is fixed to `$HOME` at create time and
-    /// never tracks the shell's `cd`s). Returns nil on any error so the
-    /// caller can fall back without crashing.
-    public func activePaneCwd(id: UUID) -> String? {
-        guard let windowIndex = bindings[id] else { return nil }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            let raw = try ctrl.displayMessage(target: target, format: "#{pane_current_path}")
-            let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-            return trimmed.isEmpty ? nil : trimmed
-        } catch {
-            reportIfTimeout(error)
-            return nil
-        }
-    }
-
-    /// Direction for `searchInScrollback`. `backward` walks toward older
-    /// output (the common case for Cmd+F on terminal history); `forward`
-    /// walks toward newer output.
-    public enum SearchDirection {
-        case backward
-        case forward
-    }
-
-    /// Enter tmux copy-mode and start a search for `query` in the
-    /// scrollback of terminal `id` (#471 gap 2). Powers the Cmd+F search
-    /// affordance. `tmux send-keys -X search-backward "<query>"` jumps the
-    /// copy-mode cursor to the most recent match; subsequent calls to
-    /// `searchAgain` step through additional matches without re-running
-    /// the search. The pane stays in copy-mode until the caller invokes
-    /// `exitCopyMode` (or the user hits ESC).
-    public func searchInScrollback(
-        id: UUID,
-        query: String,
-        direction: SearchDirection
-    ) throws {
-        guard !query.isEmpty else { return }
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            _ = try ctrl.run(["copy-mode", "-H", "-t", target])
-            let command = direction == .backward ? "search-backward" : "search-forward"
-            try ctrl.sendKeys(target: target, keys: ["-X", command, query])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Step to the next/previous match for the active search in terminal
-    /// `id`'s copy-mode (#471 gap 2). Maps to `search-again` /
-    /// `search-reverse` per tmux's own conventions.
-    public func searchAgain(id: UUID, reverse: Bool) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            let command = reverse ? "search-reverse" : "search-again"
-            try ctrl.sendKeys(target: target, keys: ["-X", command])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Leave copy-mode in terminal `id`, restoring normal shell input.
-    /// Used by the search bar's Done button (#471 gap 2) and by callers
-    /// that want to abandon a prompt-jump (#471 gap 6).
-    public func exitCopyMode(id: UUID) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            try ctrl.sendKeys(target: target, keys: ["-X", "cancel"])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Jump the copy-mode cursor to the previous OSC 133;A prompt-start
-    /// marker in terminal `id` (#471 gap 6). Requires the shell wrapper
-    /// to emit a non-passthrough OSC 133;A so tmux's emulator sees it.
-    /// Enters copy-mode if not already there.
-    public func previousPrompt(id: UUID) throws {
-        try sendPromptNav(id: id, command: "previous-prompt")
-    }
-
-    /// Sibling of `previousPrompt`. Steps forward through OSC 133;A marks.
-    public func nextPrompt(id: UUID) throws {
-        try sendPromptNav(id: id, command: "next-prompt")
-    }
-
-    private func sendPromptNav(id: UUID, command: String) throws {
-        guard let windowIndex = bindings[id] else {
-            throw TmuxBackendError.unknownTerminal(id)
-        }
-        do {
-            let ctrl = try ensureRunningServer()
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            _ = try ctrl.run(["copy-mode", "-H", "-t", target])
-            try ctrl.sendKeys(target: target, keys: ["-X", command])
-        } catch {
-            reportIfTimeout(error)
-            throw error
-        }
-    }
-
-    /// Destroy the tmux window backing `id` and forget the binding.
-    public func destroyTerminal(id: UUID) {
-        if let windowIndex = bindings[id] {
-            controller?.killWindow(index: windowIndex)
-            // CROW-1023: drop the alt-buffer latch on this path too — close-tab
-            // goes straight to `controller?.killWindow`, not `killWindow(index:)`,
-            // so the ADR's "cleared at kill" would otherwise not hold here. Prune
-            // + the register-time remove already cover a recycled index; this
-            // keeps every teardown path consistent (review).
-            observedAltBufferWindows.remove(windowIndex)
-        }
-        bindings.removeValue(forKey: id)
-        // Forget the active marker if this was the selected terminal, so a
-        // window index tmux later reuses can't be wrongly deduped away.
-        if activeTerminalID == id { activeTerminalID = nil }
-        // Cancel in-flight readiness watch Tasks so a 30s waiter doesn't
-        // fire `onReadinessChanged` against a stale id long after the tab
-        // is gone (issue #282).
-        readinessTasks.removeValue(forKey: id)?.forEach { $0.cancel() }
-        if let sentinelPath = sentinels.removeValue(forKey: id) {
-            try? FileManager.default.removeItem(atPath: sentinelPath)
-        }
-        if let logPath = wrapperLogs.removeValue(forKey: id) {
-            try? FileManager.default.removeItem(atPath: logPath)
-        }
-    }
-
-    /// Bare login shells we consider "orphaned" when a cockpit window is not
-    /// referenced by any terminal — i.e. a window left at a shell with no agent
-    /// running (#408). Anything else (claude/codex/node/an editor/…) is left
-    /// alone. tmux reports `pane_current_command` without the login-shell `-`
-    /// prefix, but we match both forms defensively.
-    nonisolated static let orphanLoginShells: Set<String> = [
-        "zsh", "-zsh", "bash", "-bash", "sh", "-sh",
-        "fish", "-fish", "dash", "-dash", "ksh", "tcsh", "csh", "login",
-    ]
-
-    /// Decide whether a single cockpit window should be reaped. Pure so the
-    /// policy is unit-testable: reap only when the window is NOT referenced by a
-    /// live terminal AND its pane is sitting at a bare login shell. Never reaps
-    /// a window running an agent (or any non-shell process), so an agent that
-    /// exited and left the user at a shell — but whose terminal still references
-    /// the window — is preserved (it's in `keep`).
-    nonisolated static func shouldReapWindow(index: Int, command: String, keep: Set<Int>) -> Bool {
-        if keep.contains(index) { return false }
-        return orphanLoginShells.contains(command)
-    }
-
-    /// Targeted-auto orphan policy (CROW-581). Reap a cockpit window that no live
-    /// terminal references (`!keep.contains`) when it is either a forgotten bare
-    /// login shell OR a positively-identified coding-agent window (its pinned
-    /// name is one of `agentWindowNames`) that has stayed orphaned across two
-    /// passes (`seenOrphanedLastPass`, the grace). NEVER reaps the session anchor
-    /// (index 0), a bound window, an unknown/infra window, or a **Manager**
-    /// (name contains "manager") — Managers are long-lived and may be unbound.
-    /// Pure so the policy is unit-testable without tmux.
-    nonisolated static func shouldReapOrphanWindow(
-        index: Int, name: String, command: String, keep: Set<Int>,
-        agentWindowNames: Set<String>, seenOrphanedLastPass: Bool
-    ) -> Bool {
-        if index == 0 { return false }
-        if keep.contains(index) { return false }
-        if name.range(of: "manager", options: [.caseInsensitive]) != nil { return false }
-        if orphanLoginShells.contains(command) { return true }
-        if agentWindowNames.contains(name) { return seenOrphanedLastPass }
-        return false
-    }
-
-    /// Live cockpit windows as (index, pinned name, foreground command). `[]` if
-    /// tmux is unavailable or the read fails.
-    public func listCockpitWindows() -> [(index: Int, name: String, command: String)] {
-        guard let ctrl = controller else { return [] }
-        do { return try ctrl.listWindows() }
-        catch { reportIfTimeout(error); return [] }
-    }
-
-    // MARK: - Scrollback health (CROW-804)
-
-    /// Pure policy: a window's scroll-up can't show the full transcript when its
-    /// pane is in the alternate buffer (no scrollback) OR its `history_limit` is
-    /// below the ceiling we bake into new windows. tmux freezes both at window
-    /// birth and can't resize/undo either in place (see `crow-tmux.conf`
-    /// history-limit caveat), so a degraded window's only remedy is recreation.
-    /// `nonisolated static` so the policy is unit-testable without tmux.
-    ///
-    /// `alternateScreenEnabled` makes the alt-buffer half of that test
-    /// KIND-AWARE (ADR-0013). Under the per-surface hybrid scroll model an
-    /// agent-TUI window deliberately runs with `alternate-screen on`, so
-    /// `alternateOn == true` there is the design working, not a stuck window —
-    /// flagging it would badge every agent tab with the ⚠ "Recreate" affordance.
-    /// The `history_limit` floor still applies to those windows, and it is what
-    /// keeps the CROW-804/#821 detection meaningful: the real pre-config
-    /// casualties measured `history_limit=5000` alongside `alternate_on=1`, so
-    /// they stay caught by the floor.
-    ///
-    /// CROW-1010 retracted the CROW-1008 inline-agent `history-limit 0` clamp,
-    /// so an agent window at 0 is a leftover of that clamp and fails the floor
-    /// like any other under-cap window (Recreate rebuilds it at 50k). A plain
-    /// shell at 0 was always degraded.
-    ///
-    /// Known blind spot: an agent window genuinely wedged in the alt buffer at
-    /// the full 50000 limit is now indistinguishable from the normal state.
-    /// That is the accepted cost of the hybrid model — the alternative is a
-    /// false ⚠ on every healthy agent surface.
-    nonisolated public static func isScrollbackDegraded(
-        historyLimit: Int,
-        alternateOn: Bool,
-        alternateScreenEnabled: Bool = false,
-        floor: Int = TmuxBackend.scrollbackHistoryLimit
-    ) -> Bool {
-        if historyLimit < floor { return true }
-        // An agent surface is SUPPOSED to be in the alt buffer.
-        return alternateScreenEnabled ? false : alternateOn
-    }
-
-    /// Pure CROW-1023 alt-buffer latch update. Given the live windows (index +
-    /// current `#{alternate_on}`) and the prior latch, return the new latch:
-    /// every window ever observed in the alt buffer, minus windows that no
-    /// longer exist.
-    ///
-    /// STICKY (union) is the point — a Claude/agent build that enters the alt
-    /// screen once keeps the capped-0 scroll model through its transient
-    /// main-buffer drops (shell-out, exit) instead of flip-flopping xterm's
-    /// scrollback every poll. PRUNE (intersect with live indices) keeps a killed
-    /// window's observation from leaking to whatever tmux later assigns that
-    /// index. `nonisolated static` so the policy is unit-testable without tmux,
-    /// exactly like `isScrollbackDegraded`.
-    nonisolated public static func updatedAltBufferLatch(
-        liveWindows: [(index: Int, alternateOn: Bool)],
-        prior: Set<Int>
-    ) -> Set<Int> {
-        var latch = prior
-        for w in liveWindows where w.alternateOn { latch.insert(w.index) }
-        latch.formIntersection(liveWindows.map(\.index))
-        return latch
-    }
-
-    /// The per-window classifications the web UI needs, from ONE `list-windows`
-    /// read: which windows are scrollback-degraded (CROW-804 ⚠ Recreate), which
-    /// run the agent-TUI scroll model (ADR-0013 wheel/mouse routing), and which
-    /// have actually entered the alternate buffer (CROW-1023, the sticky
-    /// `uses_alternate_screen` latch).
-    ///
-    /// They ship together on every `list-terminals` RPC, and each is derived
-    /// from the same three fields, so reading twice would fork a second `tmux`
-    /// subprocess per call for nothing.
-    ///
-    /// Returns `nil` when tmux is unavailable or the read fails — deliberately
-    /// NOT a triple of empty sets. Empty is a perfectly valid SUCCESS (a server
-    /// of nothing but plain shells), so emptiness cannot double as a failure
-    /// signal. Callers that need to fall back to a different source of truth on
-    /// failure — `list-terminals` re-deriving `agent_surface` from
-    /// `SessionTerminal.isAgentSurface`, and `uses_alternate_screen` from the
-    /// `AgentRegistry` capability — can only do that if failure is
-    /// distinguishable. Callers that are happy to fail open collapse it with
-    /// `?? []`.
-    ///
-    /// Not read-only: it maintains the alt-buffer latch (`observedAltBufferWindows`).
-    /// Confined to `@MainActor` like the rest of this type, so the mutation is
-    /// race-free.
-    public func windowScrollbackClassification(
-        floor: Int = TmuxBackend.scrollbackHistoryLimit
-    ) -> (degraded: Set<Int>, agentSurfaces: Set<Int>, altBuffer: Set<Int>)? {
-        guard let ctrl = controller else { return nil }
-        do {
-            let windows = try ctrl.listWindowScrollback()
-            let degraded = windows.filter {
-                Self.isScrollbackDegraded(
-                    historyLimit: $0.historyLimit,
-                    alternateOn: $0.alternateOn,
-                    alternateScreenEnabled: $0.alternateScreenEnabled,
-                    floor: floor)
-            }
-            let agents = windows.filter(\.alternateScreenEnabled)
-            // CROW-1023: update the sticky alt-buffer latch from this read (pure
-            // policy in `updatedAltBufferLatch`), then report it as the third set.
-            observedAltBufferWindows = Self.updatedAltBufferLatch(
-                liveWindows: windows.map { ($0.index, $0.alternateOn) },
-                prior: observedAltBufferWindows)
-            return (Set(degraded.map(\.index)),
-                    Set(agents.map(\.index)),
-                    observedAltBufferWindows)
-        } catch {
-            reportIfTimeout(error)
-            return nil
-        }
-    }
-
-    /// Window indices whose scrollback is degraded per `isScrollbackDegraded`.
-    /// Fails open (`[]`) — not badging a window on a failed read is the safe
-    /// direction, and matches `listCockpitWindows`. Callers needing BOTH this
-    /// and the agent-surface set should use `windowScrollbackClassification`
-    /// so tmux is only read once.
-    public func degradedWindowIndices(floor: Int = TmuxBackend.scrollbackHistoryLimit) -> Set<Int> {
-        windowScrollbackClassification(floor: floor)?.degraded ?? []
-    }
-
-    /// Window indices configured as agent-TUI surfaces (`alternate-screen on`),
-    /// i.e. the windows that own their own viewport + scrollback under the
-    /// hybrid scroll model (ADR-0013). Read from tmux rather than inferred from
-    /// window names so the daemon and the web client route on the SAME ground
-    /// truth the daemon actually applied.
-    public func agentSurfaceWindowIndices() -> Set<Int> {
-        windowScrollbackClassification()?.agentSurfaces ?? []
-    }
-
-    /// Window indices that have entered the alternate buffer at least once and
-    /// are latched to the capped-0 scroll model (CROW-1023). This — not the
-    /// static per-kind capability — is what `list-terminals` forwards as
-    /// `uses_alternate_screen`. Fails open (`[]`); a window not yet observed in
-    /// the alt buffer is treated as inline (unified 50k), which is the safe
-    /// direction (a real alt-buffer build re-latches within a poll, an inline
-    /// build stays scrollable).
-    public func altBufferWindowIndices() -> Set<Int> {
-        windowScrollbackClassification()?.altBuffer ?? []
-    }
-
-    /// Give one window the agent-TUI scroll model: `alternate-screen on`, so a
-    /// repainting agent keeps its frames in the alt buffer (which has no
-    /// scrollback) instead of depositing every repaint into the shared 50k
-    /// history as duplicate-frame sediment (#822, ADR-0013).
-    ///
-    /// Best-effort by design: the window is already usable without it, so a
-    /// failure here must never fail terminal creation. Returns whether it stuck.
-    @discardableResult
-    public func enableAlternateScreen(index: Int) -> Bool {
-        guard let ctrl = controller else { return false }
-        do {
-            try ctrl.setWindowOption(index: index, name: "alternate-screen", value: "on")
-            return true
-        } catch {
-            reportIfTimeout(error)
-            CrowLog.info("[Crow] could not set alternate-screen on window \(index): \(error)")
-            return false
-        }
-    }
-
-    /// Live per-window scrollback tuple. Empty when tmux is down. Test/debug
-    /// counterpart of `windowScrollbackClassification` that keeps the raw
-    /// `history_limit` so CROW-1010 can assert inline agents keep the 50k cap.
-    func windowScrollbackSnapshot() -> [(index: Int, historyLimit: Int, alternateOn: Bool, alternateScreenEnabled: Bool)] {
-        guard let ctrl = controller else { return [] }
-        return (try? ctrl.listWindowScrollback()) ?? []
-    }
-
-    /// Kill the cockpit window at `index`. Passthrough to the controller so
-    /// callers outside `TmuxBackend` (e.g. the CROW-804 terminal recreate in
-    /// `SessionService`) can drop a degraded window before re-registering a
-    /// fresh one. No-op when tmux is unavailable.
-    public func killWindow(index: Int) {
-        controller?.killWindow(index: index)
-        // CROW-1023: drop the alt-buffer latch for a killed index up front, so a
-        // recreate that reuses it starts inline until it re-enters the alt screen.
-        observedAltBufferWindows.remove(index)
-    }
-
-    /// Reap orphaned cockpit windows per `shouldReapOrphanWindow` (targeted-auto).
-    /// `keepWindowIndices` are windows referenced by persisted terminals — unioned
-    /// with the in-memory `bindings` so a just-adopted window is never reaped.
-    /// `agentWindowNames` are the display names `new-terminal` pins on managed
-    /// agent windows. Tracks the agent-orphan set for the next pass's grace.
-    /// Best-effort; returns the count reaped (CROW-581).
-    @discardableResult
-    public func reconcileOrphanWindows(keepWindowIndices: Set<Int>, agentWindowNames: Set<String>) -> Int {
-        guard let ctrl = controller else { return 0 }
-        let keep = keepWindowIndices.union(bindings.values)
-        let windows = listCockpitWindows()
-        let previouslyOrphaned = orphanGraceWindows
-        var stillOrphanedAgents: Set<Int> = []
-        var reaped = 0
-        for w in windows {
-            // Agent-named orphans are grace candidates for the next pass.
-            if w.index != 0, !keep.contains(w.index), agentWindowNames.contains(w.name),
-               w.name.range(of: "manager", options: [.caseInsensitive]) == nil {
-                stillOrphanedAgents.insert(w.index)
-            }
-            if Self.shouldReapOrphanWindow(
-                index: w.index, name: w.name, command: w.command, keep: keep,
-                agentWindowNames: agentWindowNames,
-                seenOrphanedLastPass: previouslyOrphaned.contains(w.index)) {
-                ctrl.killWindow(index: w.index)
-                CrowLog.info("[CrowTelemetry tmux:orphan_window_reaped index=\(w.index) name=\(w.name) command=\(w.command)]")
-                reaped += 1
-            }
-        }
-        orphanGraceWindows = stillOrphanedAgents
-        if reaped > 0 { CrowLog.info("[Crow] Reaped \(reaped) orphaned cockpit window(s) (CROW-581)") }
-        return reaped
-    }
-
-    /// Decide whether the Manager agent has exited, from one poll sample of its
-    /// window's foreground command (#558). Because the agent (`claude …`) runs
-    /// *inside* the window's shell wrapper rather than as the pane's direct
-    /// child, its exit doesn't kill the pane — the foreground just falls back to
-    /// a bare login shell. So we report an exit only once we've seen the agent
-    /// actually running (`sawAgentRunning`, a non-shell foreground) and now see
-    /// a bare login shell. `nil` (window gone / read failed) is never an exit —
-    /// that path also covers teardown, keeping restart/shutdown false-positive
-    /// free. Pure so the transition policy is unit-testable without tmux.
-    nonisolated static func managerAgentDidExit(paneCommand: String?, sawAgentRunning: Bool) -> Bool {
-        guard sawAgentRunning, let command = paneCommand else { return false }
-        return orphanLoginShells.contains(command)
-    }
-
-    /// Advance the exit-monitor state machine by one poll sample, so the whole
-    /// transition — not just its terminal condition — is unit-testable without
-    /// tmux (#558). `sample` is the window's foreground command, or `nil` when
-    /// the read was inconclusive (binding absent / `display-message` threw /
-    /// window gone). Returns the updated `sawAgentRunning` latch and whether an
-    /// exit should fire. A non-empty non-shell command latches "agent running";
-    /// a bare login shell after that latch fires; `nil` and `""` are no-ops.
-    nonisolated static func advanceExitMonitor(
-        sawAgentRunning: Bool, sample: String?
-    ) -> (sawAgentRunning: Bool, fired: Bool) {
-        guard let command = sample else { return (sawAgentRunning, false) }
-        if orphanLoginShells.contains(command) {
-            return (sawAgentRunning, managerAgentDidExit(paneCommand: command, sawAgentRunning: sawAgentRunning))
-        }
-        if !command.isEmpty { return (true, false) }
-        return (sawAgentRunning, false)
-    }
-
-    /// Reap cockpit windows that no live terminal references AND that are
-    /// sitting at a bare login shell — leaked windows from a timed-out
-    /// `new-window` or a forgotten terminal (#408). `keepWindowIndices` is the
-    /// set of window indices referenced by persisted terminals; it is unioned
-    /// with the in-memory `bindings` so a window created/adopted this run is
-    /// never reaped. Best-effort; returns the count reaped.
-    @discardableResult
-    public func reapUnboundCockpitWindows(keepWindowIndices: Set<Int>) -> Int {
-        guard let ctrl = controller else { return 0 }
-        let keep = keepWindowIndices.union(bindings.values)
-        let windows: [(index: Int, command: String)]
-        do {
-            windows = try ctrl.listWindowCommands()
-        } catch {
-            reportIfTimeout(error)
-            return 0
-        }
-        var reaped = 0
-        for window in windows where Self.shouldReapWindow(index: window.index, command: window.command, keep: keep) {
-            ctrl.killWindow(index: window.index)
-            CrowLog.info("[CrowTelemetry tmux:orphan_window_reaped index=\(window.index) command=\(window.command)]")
-            reaped += 1
-        }
-        if reaped > 0 {
-            CrowLog.info("[Crow] Reaped \(reaped) orphaned bare-shell cockpit window(s) (#408)")
-        }
-        return reaped
-    }
-
-    // MARK: - Manager exit monitor (#558)
-
-    /// Watch the Manager terminal's tmux window and fire `onExit` the first time
-    /// its foreground command falls back to a bare login shell after the agent
-    /// was seen running — i.e. the Manager's `claude`/`codex`/… process exited.
-    ///
-    /// A `tmux attach-session` client only signals when the whole client dies,
-    /// so it can't tell a per-window agent exit apart (#558). We poll
-    /// `#{pane_current_command}` for the Manager window instead — the same
-    /// signal the orphan reaper reads.
-    ///
-    /// At most one monitor runs; a second call cancels the first. The poll skips
-    /// samples where the binding is absent (async adopt on launch hasn't landed
-    /// yet) or the `display-message` throws (window gone / server down), so
-    /// attach-client teardown and restart/shutdown never false-positive. Stops
-    /// after firing once; `SessionService` re-arms it on the next Manager launch.
-    ///
-    /// The blocking `display-message` runs off the main actor (`Task.detached`)
-    /// and only the `onExit` hop-back touches the UI thread — a perpetual poll
-    /// must never stall AppKit even if the tmux subprocess wedges to its
-    /// watchdog timeout.
-    ///
-    /// Sampling floor: an agent that both launches and exits inside one
-    /// `pollInterval` is never observed running, so `sawAgentRunning` stays
-    /// false and no banner fires. That's an accepted missed-detection inherent
-    /// to polling, not a false positive.
-    public func startManagerExitMonitor(
-        id: UUID,
-        pollInterval: TimeInterval = 3.0,
-        onExit: @escaping @MainActor () -> Void
-    ) {
-        stopManagerExitMonitor()
-        let nanos = UInt64(pollInterval * 1_000_000_000)
-        managerExitMonitor = Task { [weak self] in
-            var sawAgentRunning = false
-            while !Task.isCancelled {
-                try? await Task.sleep(nanoseconds: nanos)
-                if Task.isCancelled { return }
-                guard let self else { return }
-                // Capture the tmux handle + window index on the main actor, then
-                // read `#{pane_current_command}` off it: `TmuxController.run`
-                // blocks the calling thread (on a semaphore since #653, without
-                // pumping the run loop), so keep that subprocess off the UI
-                // thread. A missing binding/controller yields a `nil` sample (skip).
-                let ctrl = self.controller
-                let windowIndex = self.bindings[id]
-                let raw: String? = await Task.detached { () -> String? in
-                    guard let ctrl, let windowIndex else { return nil }
-                    return try? ctrl.displayMessage(
-                        target: "\(ctrl.sessionName):\(windowIndex)",
-                        format: "#{pane_current_command}"
-                    )
-                }.value
-                let sample = raw?.trimmingCharacters(in: .whitespacesAndNewlines)
-                if Task.isCancelled { return }
-                let (nextSaw, fired) = Self.advanceExitMonitor(sawAgentRunning: sawAgentRunning, sample: sample)
-                sawAgentRunning = nextSaw
-                if fired {
-                    CrowLog.info("[CrowTelemetry manager:exit_detected terminal=\(id) command=\(sample ?? "")]")
-                    onExit()
-                    return
-                }
-            }
-        }
-    }
-
-    /// Cancel the Manager exit monitor if one is running. Idempotent.
-    public func stopManagerExitMonitor() {
-        managerExitMonitor?.cancel()
-        managerExitMonitor = nil
-    }
-
     /// Reset the cockpit's active-window marker so the next `makeActive` runs a
     /// real `select-window` instead of short-circuiting. Called on the light
     /// "attach client died, server still alive" recovery path (#588) — the
@@ -1119,473 +260,10 @@ public final class TmuxBackend {
         bindings[id] != nil
     }
 
-    // MARK: - Internal helpers
-
-    private func ensureRunningServer() throws -> TmuxController {
-        try ensureRunningServerReportingResurrection().ctrl
-    }
-
-    /// Like `ensureRunningServer()`, but also reports whether the cockpit
-    /// session had to be created from scratch (`resurrected == true`) —
-    /// i.e. the server was NOT running when the caller needed it. Callers
-    /// that see a persisted binding fail against a resurrected server know
-    /// the whole server died, not just one window (#588).
-    private func ensureRunningServerReportingResurrection()
-        throws -> (ctrl: TmuxController, resurrected: Bool)
-    {
-        if let ctrl = controller {
-            if ctrl.hasSession() { return (ctrl, false) }
-            // We had a live cockpit this run and it's gone — the server died
-            // mid-run (or is hung past the watchdog; recovery handles both
-            // identically). Fire BEFORE resurrecting so the handler can
-            // observe the dead state; it's reentrancy-guarded and hops to a
-            // later main-actor turn, so recovery never races this call.
-            CrowLog.info("[CrowTelemetry tmux:server_died_midrun bindings=\(bindings.count)]")
-            onServerLost?()
-        }
-        guard !tmuxBinary.isEmpty, !socketPath.isEmpty else {
-            // Backend wasn't configured this run (tmux not discovered).
-            // Throw rather than precondition-crash — callers catch and surface
-            // an error overlay.
-            throw TmuxBackendError.notConfigured
-        }
-        let ctrl = TmuxController(
-            tmuxBinary: tmuxBinary,
-            socketPath: socketPath,
-            sessionName: TmuxBackend.cockpitSessionName
-        )
-        guard let confURL = BundledResources.tmuxConfURL else {
-            throw TmuxBackendError.bundledResourceMissing("crow-tmux.conf")
-        }
-        // The cockpit session may already be live from a prior Crow launch
-        // (#330 stable socket). If so, the bundled conf the server loaded at
-        // `-f` time may now be stale relative to the file on disk (#450) —
-        // capture the pre-attach state so we can reconcile below.
-        let serverWasAlreadyLive = ctrl.hasSession()
-        try Self.ensureCockpitSession(ctrl, configPath: confURL.path)
-        controller = ctrl
-        if serverWasAlreadyLive {
-            Self.reconcileBundledConfigIfStale(controller: ctrl, configURL: confURL)
-        }
-        return (ctrl, !serverWasAlreadyLive)
-    }
-
-    // MARK: - Stale-config reconciliation (#450)
-
-    /// Re-source the bundled tmux conf on a live server iff the file on disk
-    /// has been modified since the server started. Non-destructive: existing
-    /// windows/sessions survive a `source-file` — server-scoped options
-    /// (mouse, status, escape-time, …) update in place. Failures are logged
-    /// and swallowed; a stale conf is not worth aborting startup over.
-    ///
-    /// Caveat: the bundled conf includes `set -gas terminal-features ',…'`
-    /// which re-appends on each source-file. tmux tolerates duplicate feature
-    /// flags (merged by name) so the duplication is benign.
-    nonisolated static func reconcileBundledConfigIfStale(
-        controller: TmuxController,
-        configURL: URL
-    ) {
-        let confPath = configURL.path
-        let confMTime = (try? FileManager.default.attributesOfItem(atPath: confPath))?[.modificationDate] as? Date
-        let serverStart = serverStartTime(controller: controller)
-
-        guard shouldReconcile(configMTime: confMTime, serverStartTime: serverStart) else {
-            CrowLog.info("[CrowTelemetry tmux:config_reconcile_skipped reason=fresh]")
-            return
-        }
-
-        do {
-            try controller.run(["source-file", confPath])
-            CrowLog.info("[CrowTelemetry tmux:config_reconciled path=\(confPath)]")
-        } catch {
-            CrowLog.info("[CrowTelemetry tmux:config_reconcile_failed error=\"\(error)\"]")
-        }
-    }
-
-    /// Re-source the bundled `crow-tmux.conf` against the live tmux server
-    /// unconditionally — driven by the "Reload Terminal Config" menu item
-    /// (#475), where the user has explicitly asked for a reload. Unlike
-    /// `reconcileBundledConfigIfStale`, this skips the mtime gate.
-    ///
-    /// Returns `nil` on success, or a human-readable error string the caller
-    /// can surface in a banner. Idempotent: `source-file` against a live
-    /// server updates server-scoped options in place; existing windows and
-    /// sessions are unaffected.
-    ///
-    /// `async` and off the main actor: both `hasSession()` and `run()` block the
-    /// calling thread, and the RPC handler invokes this from `MainActor.run`, so
-    /// on the main actor they stall every other MainActor-bound RPC behind them
-    /// (CROW-874). `TmuxController` is a `Sendable` struct of three strings, so
-    /// it crosses cleanly — the same shape `startManagerExitMonitor` uses.
-    ///
-    /// This is one of ~24 `@MainActor` entry points on this type that call
-    /// `TmuxController` synchronously; bounding `run()` itself is what fixes the
-    /// class. Moving the rest needs `sendText`/`makeActive` and friends to stop
-    /// being sync `throws` APIs, which touches every caller.
-    public func reloadBundledConfig() async -> String? {
-        guard let ctrl = controller else {
-            return "tmux server is not running"
-        }
-        guard let confURL = BundledResources.tmuxConfURL else {
-            return "bundled crow-tmux.conf not found"
-        }
-        return await Task.detached {
-            guard ctrl.hasSession() else {
-                return "tmux server is not running"
-            }
-            do {
-                try ctrl.run(["source-file", confURL.path])
-                CrowLog.info("[CrowTelemetry tmux:config_reloaded_by_user path=\(confURL.path)]")
-                return nil
-            } catch {
-                CrowLog.info("[CrowTelemetry tmux:config_reload_failed error=\"\(error)\"]")
-                return "\(error)"
-            }
-        }.value
-    }
-
-    /// Pure policy: reconcile when either timestamp is missing (conservative
-    /// — a redundant `source-file` is cheap) or when the conf is newer than
-    /// the running server.
-    nonisolated static func shouldReconcile(configMTime: Date?, serverStartTime: Date?) -> Bool {
-        guard let configMTime, let serverStartTime else { return true }
-        return configMTime > serverStartTime
-    }
-
-    /// `tmux display -p '#{start_time}'` → Unix epoch as a string. Returns
-    /// nil on any IO/parse failure; callers treat nil as "unknown — reconcile
-    /// to be safe".
-    nonisolated static func serverStartTime(controller: TmuxController) -> Date? {
-        guard let raw = try? controller.run(["display", "-p", "#{start_time}"]) else { return nil }
-        let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard let epoch = TimeInterval(trimmed) else { return nil }
-        return Date(timeIntervalSince1970: epoch)
-    }
-
-    /// Ensure the cockpit session is live, adopting an existing one if a
-    /// concurrent caller won the `new-session` race.
-    ///
-    /// **The adopt branch below is load-bearing — do not delete it.** Its
-    /// original justification was in-process: `TmuxController.run` blocked on
-    /// `Process.waitUntilExit()`, which pumps the main run loop, so a
-    /// `new-session` could be re-entered by another `ensureRunningServer()`
-    /// caller before `controller` was cached. That mechanism is gone — since
-    /// #653 the wait is a semaphore that does not pump — but the race is not.
-    ///
-    /// It is now a **cross-process** TOCTOU, which no in-process argument can
-    /// close: `TerminalCockpit.ensureSession` in `crowd` performs the identical
-    /// `hasSession()` → `newSessionDetached` against the *same* socket
-    /// (`appTmuxSocketPath()`, #330), and runs it on every daemon startup.
-    /// Whoever wins creates `crow-cockpit`; the rest must ADOPT it, not
-    /// re-create it — `new-session` errors with "duplicate session", and
-    /// because that throws, the loser never cached `controller`, so every
-    /// subsequent call kept failing and every terminal rendered blank (#326).
-    ///
-    /// `nonisolated static` so the adopt branch is testable without a real
-    /// tmux server or the main actor — it touches no instance/actor state, and
-    /// nothing pins it to the main actor.
-    nonisolated static func ensureCockpitSession(
-        _ ctrl: CockpitSessionStarter,
-        configPath: String?,
-        // The "session anchor" is a no-op long-running command — kept alive so
-        // the session persists even if every window is closed by the user.
-        // /usr/bin/tail -f /dev/null is the conventional choice.
-        anchorCommand: String = "/usr/bin/tail -f /dev/null"
-    ) throws {
-        if ctrl.hasSession() { return }
-        do {
-            try ctrl.newSessionDetached(configPath: configPath, env: [:], command: anchorCommand)
-        } catch {
-            // Lost the creation race after the `hasSession()` check above: a
-            // reentrant caller created the session while our `new-session`
-            // subprocess was starting. The session exists, which is exactly
-            // the post-condition we want — adopt it rather than propagating
-            // the spurious "duplicate session" failure.
-            guard ctrl.hasSession() else { throw error }
-        }
-    }
-
-    private func sentinelPath(for id: UUID) -> String {
-        let dir = ProcessInfo.processInfo.environment["TMPDIR"] ?? "/tmp/"
-        return (dir as NSString)
-            .appendingPathComponent("crow-ready-\(id.uuidString).sentinel")
-    }
-
-    /// Per-terminal log path for `crow-shell-wrapper.sh` stage breadcrumbs
-    /// (issue #256). Stable across `registerTerminal` / `adoptTerminal` /
-    /// `retryReadinessWatch` for a given terminal UUID.
-    private func wrapperLogPath(for id: UUID) -> String {
-        let dir = ProcessInfo.processInfo.environment["TMPDIR"] ?? "/tmp/"
-        return (dir as NSString)
-            .appendingPathComponent("crow-wrapper-\(id.uuidString).log")
-    }
-
-    private func startReadinessWatch(id: UUID, sentinelPath: String) {
-        startReadinessWatch(id: id, sentinelPath: sentinelPath, timeoutBudget: 30.0)
-    }
-
-    private func startReadinessWatch(id: UUID, sentinelPath: String, timeoutBudget: TimeInterval) {
-        let waiter = SentinelWaiter()
-        // Periodic progress beacon every 10s so operators tailing the log can
-        // see the watch is alive and whether the sentinel has appeared yet
-        // (issue #256). Cancelled when the waiter resolves.
-        let progressTask = Task { [weak self] in
-            let startedAt = Date()
-            while !Task.isCancelled {
-                try? await Task.sleep(nanoseconds: 10_000_000_000)
-                if Task.isCancelled { break }
-                let elapsed = Int(Date().timeIntervalSince(startedAt) * 1000)
-                let exists = FileManager.default.fileExists(atPath: sentinelPath)
-                _ = self  // keep the capture; method-level NSLog is fine
-                CrowLog.info("[CrowTelemetry tmux:first_prompt_progress terminal=\(id) elapsed_ms=\(elapsed) sentinel_exists=\(exists)]")
-            }
-        }
-        let waiterTask = Task { [weak self] in
-            // 30s default budget (was 5s). On app restart with many managed
-            // terminals hydrating concurrently, shell startup is CPU-contended
-            // and the wrapper's first precmd may not fire within 5s. Callers
-            // can pass a longer budget for retries (see retryReadinessWatch).
-            let elapsed = await waiter.waitForPrompt(
-                sentinelPath: sentinelPath,
-                timeout: timeoutBudget
-            )
-            progressTask.cancel()
-            await MainActor.run { [weak self] in
-                // Bail if the terminal was destroyed (or the backend went
-                // away) while we were waiting. Without this guard a 30s
-                // waiter could fire readiness for a tab the user closed
-                // 29s ago — issue #282.
-                guard let self, self.bindings[id] != nil else { return }
-                if let elapsed {
-                    let ms = Int(elapsed * 1000)
-                    CrowLog.info("[CrowTelemetry tmux:first_prompt_ms=\(ms) terminal=\(id)]")
-                    self.onReadinessChanged?(id, .shellReady)
-                } else {
-                    // Genuine timeout. Most likely the shell is alive but its
-                    // startup is pathologically slow (heavy zshrc + concurrent
-                    // hydrate, cold tmux server + App Nap on a backgrounded
-                    // app); less likely the wrapper failed to install the
-                    // precmd hook (exotic shell) or the shell crashed at start.
-                    //
-                    // Surface this via `.timedOut` rather than lying about
-                    // readiness. Auto-paste of the launch command relies on a
-                    // live `zle`, so pasting blind here can leave the pane in
-                    // an unrecoverable state (visible command, no Claude TUI,
-                    // bytes consumed by half-initialized subshells). The UI
-                    // renders a Retry affordance for `.timedOut`; the
-                    // `didBecomeActive` observer also re-arms automatically
-                    // when the app returns to the foreground.
-                    let ms = Int(timeoutBudget * 1000)
-                    CrowLog.info("[CrowTelemetry tmux:first_prompt_timeout terminal=\(id) budget_ms=\(ms)]")
-                    // Capture stage-by-stage diagnostics and dump them to the
-                    // system log alongside the timeout marker. The UI surfaces
-                    // the same bundle via "Copy diagnostics" (issue #256).
-                    let bundle = self.captureDiagnostics(id: id)
-                    CrowLog.info("[CrowTelemetry tmux:first_prompt_diagnostics terminal=\(id)]\n\(bundle)")
-                    self.onReadinessChanged?(id, .timedOut)
-                }
-            }
-        }
-        // Track both Tasks so `destroyTerminal` can cancel them (#282).
-        // `retryReadinessWatch` may call us again for the same id; the
-        // previous Tasks are already resolved (or about to be), so appending
-        // here rather than replacing keeps the contract simple.
-        readinessTasks[id, default: []].append(contentsOf: [progressTask, waiterTask])
-    }
-
-    /// Re-arm the readiness watch for a terminal whose first attempt timed
-    /// out. Clears the stale sentinel file (in case the wrapper now writes
-    /// to it asynchronously) and starts a fresh watch with a longer budget.
-    /// Safe to call repeatedly; previous watches resolve independently.
-    public func retryReadinessWatch(id: UUID, timeoutBudget: TimeInterval = 120.0) {
-        guard let sentinelPath = sentinels[id] else { return }
-        try? FileManager.default.removeItem(atPath: sentinelPath)
-        startReadinessWatch(id: id, sentinelPath: sentinelPath, timeoutBudget: timeoutBudget)
-    }
-
-    /// Build a stage-by-stage diagnostic bundle for terminal `id`. Captures
-    /// pane contents, pane PID + process tree, sentinel state, and the
-    /// wrapper's breadcrumb log. Each section is wrapped so a single missing
-    /// piece doesn't lose the rest; per-section output is capped to keep the
-    /// clipboard payload sane. Called from the readiness-watch timeout path
-    /// (logged via NSLog) and from the UI "Copy diagnostics" button
-    /// (issue #256).
-    public func captureDiagnostics(id: UUID) -> String {
-        let sectionCap = 8_192
-        var lines: [String] = []
-        lines.append("=== Crow tmux readiness diagnostics ===")
-        lines.append("terminal=\(id)")
-        lines.append("captured_at=\(ISO8601DateFormatter().string(from: Date()))")
-        lines.append("")
-
-        // Section 1: environment & host
-        lines.append("--- environment ---")
-        let env = ProcessInfo.processInfo.environment
-        lines.append("SHELL=\(env["SHELL"] ?? "")")
-        lines.append("PATH=\(env["PATH"] ?? "")")
-        lines.append("TERM=\(env["TERM"] ?? "")")
-        lines.append("USER=\(env["USER"] ?? "")")
-        if !tmuxBinary.isEmpty,
-           let ver = TmuxController.versionString(tmuxBinary: tmuxBinary) {
-            lines.append("tmux=\(ver)")
-        } else {
-            lines.append("tmux=<unknown>")
-        }
-        if let dscl = runShortCommand(
-            "/usr/bin/dscl",
-            ["." , "-read", "/Users/\(env["USER"] ?? "")", "UserShell"]
-        ) {
-            lines.append("dscl=\(dscl.trimmingCharacters(in: .whitespacesAndNewlines))")
-        }
-        lines.append("")
-
-        // Section 2: tmux state for this terminal's window
-        lines.append("--- tmux state ---")
-        guard let windowIndex = bindings[id] else {
-            lines.append("no binding for terminal \(id) — window never created")
-            lines.append("")
-            return appendSentinelAndLog(id: id, sectionCap: sectionCap, lines: lines)
-        }
-        let ctrl = controller
-        if let ctrl {
-            let target = "\(ctrl.sessionName):\(windowIndex)"
-            lines.append("target=\(target)")
-
-            // pane_pid + pane_current_command — what's actually running in
-            // the pane right now.
-            var panePID: Int32?
-            if let info = try? ctrl.displayMessage(
-                target: target,
-                format: "#{pane_pid} #{pane_current_command}"
-            ) {
-                let trimmed = info.trimmingCharacters(in: .whitespacesAndNewlines)
-                lines.append("display_message=\(trimmed)")
-                if let firstField = trimmed.split(separator: " ").first,
-                   let pid = Int32(firstField) {
-                    panePID = pid
-                }
-            } else {
-                lines.append("display_message=<failed>")
-            }
-            lines.append("")
-
-            // ps on the pane PID + immediate descendants. Reveals whether the
-            // wrapper is still alive or has exec'd into the shell.
-            lines.append("--- process tree ---")
-            if let pid = panePID {
-                if let ps = runShortCommand(
-                    "/bin/ps",
-                    ["-o", "pid,ppid,stat,etime,command", "-p", "\(pid)"]
-                ) {
-                    lines.append(ps.trimmingCharacters(in: .whitespacesAndNewlines))
-                }
-                if let children = runShortCommand("/usr/bin/pgrep", ["-P", "\(pid)"]) {
-                    let childPIDs = children.split(separator: "\n").map(String.init).filter { !$0.isEmpty }
-                    for child in childPIDs {
-                        if let ps = runShortCommand(
-                            "/bin/ps",
-                            ["-o", "pid,ppid,stat,etime,command", "-p", child]
-                        ) {
-                            lines.append(ps.trimmingCharacters(in: .whitespacesAndNewlines))
-                        }
-                    }
-                }
-            } else {
-                lines.append("<no pane pid available>")
-            }
-            lines.append("")
-
-            // Pane capture — usually the single most useful signal: shows
-            // whether we're stuck at the shell prompt, mid-.zshrc, or showing
-            // a python traceback from oh-my-zsh.
-            lines.append("--- pane capture (last 200 lines) ---")
-            if let pane = try? ctrl.capturePane(target: target, linesBack: 200) {
-                lines.append(truncated(pane, max: sectionCap))
-            } else {
-                lines.append("<capture-pane failed>")
-            }
-            lines.append("")
-        } else {
-            lines.append("controller not initialized")
-            lines.append("")
-        }
-
-        return appendSentinelAndLog(id: id, sectionCap: sectionCap, lines: lines)
-    }
-
-    private func appendSentinelAndLog(id: UUID, sectionCap: Int, lines: [String]) -> String {
-        var out = lines
-
-        // Sentinel state — exists? size? parent writable?
-        out.append("--- sentinel ---")
-        if let path = sentinels[id] {
-            out.append("path=\(path)")
-            let fm = FileManager.default
-            let exists = fm.fileExists(atPath: path)
-            out.append("exists=\(exists)")
-            if exists, let attrs = try? fm.attributesOfItem(atPath: path) {
-                if let size = attrs[.size] as? Int { out.append("size=\(size)") }
-                if let mtime = attrs[.modificationDate] as? Date {
-                    out.append("mtime=\(ISO8601DateFormatter().string(from: mtime))")
-                }
-            }
-            let parent = (path as NSString).deletingLastPathComponent
-            out.append("parent=\(parent) parent_writable=\(fm.isWritableFile(atPath: parent))")
-        } else {
-            out.append("no sentinel path recorded for terminal \(id)")
-        }
-        out.append("")
-
-        // Wrapper log — the breadcrumb trail.
-        out.append("--- wrapper log ---")
-        if let path = wrapperLogs[id] {
-            out.append("path=\(path)")
-            if let data = try? String(contentsOfFile: path, encoding: .utf8) {
-                out.append(truncated(data, max: sectionCap))
-            } else {
-                out.append("<log not readable or absent>")
-            }
-        } else {
-            out.append("no wrapper log path recorded for terminal \(id)")
-        }
-        out.append("")
-        out.append("=== end diagnostics ===")
-        return out.joined(separator: "\n")
-    }
-
-    /// Run a short command and return its stdout (≤2s timeout). Used by
-    /// `captureDiagnostics` so any single subprocess hanging can't wedge the
-    /// main actor. Returns `nil` on any failure (missing binary, non-zero
-    /// exit, timeout) so the caller can fall back gracefully.
-    private func runShortCommand(_ launchPath: String, _ args: [String]) -> String? {
-        guard FileManager.default.isExecutableFile(atPath: launchPath) else { return nil }
-        let p = Process()
-        p.executableURL = URL(fileURLWithPath: launchPath)
-        p.arguments = args
-        let outPipe = Pipe()
-        p.standardOutput = outPipe
-        p.standardError = Pipe()
-        do { try p.run() } catch { return nil }
-        let queue = DispatchQueue.global(qos: .utility)
-        let killer = DispatchWorkItem { if p.isRunning { p.terminate() } }
-        queue.asyncAfter(deadline: .now() + 2.0, execute: killer)
-        p.waitUntilExit()
-        killer.cancel()
-        guard p.terminationStatus == 0 else { return nil }
-        let data = outPipe.fileHandleForReading.readDataToEndOfFile()
-        return String(data: data, encoding: .utf8)
-    }
-
-    private func truncated(_ s: String, max: Int) -> String {
-        if s.count <= max { return s }
-        let head = s.prefix(max)
-        return head + "\n…(truncated; \(s.count - max) chars omitted)"
-    }
-
     /// Forward .timedOut errors to the unresponsive callback. Other errors
     /// pass through silently — they're regular CLI failures the caller
     /// already handles.
-    private func reportIfTimeout(_ error: Error) {
+    func reportIfTimeout(_ error: Error) {
         if let tmuxError = error as? TmuxError, case .timedOut = tmuxError {
             CrowLog.info("[CrowTelemetry tmux:server_unresponsive error=\"\(tmuxError)\"]")
             onUnresponsive?(tmuxError)
